SquareUp comms/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlmodel>=0.0.22
asyncpg>=0.30.0
alembic>=1.14.0
anthropic>=0.42.0
firebase-admin>=6.6.0
python-multipart>=0.0.18
cryptography>=44.0.0
livekit-api>=0.7.0
pydantic-settings>=2.7.0
python-jose>=3.3.0
httpx>=0.28.0
openai>=1.0.0
google-api-python-client>=2.150.0
google-auth-oauthlib>=1.2.0
google-auth>=2.36.0
aiosqlite>=0.20.0
eval-type-backport>=0.2.0
slowapi>=0.1.9
